Refactor code into smaller files with better dependencies
No functional changes.
