+++ examples/low_level_dataset.rs
@@ -1,9 +1,12 @@
 use std::{collections::HashMap, iter, num::NonZeroU64, sync::Arc};
 
 use icechunk::{
+    dataset::{
+        ChunkIndices, ChunkKeyEncoding, ChunkPayload, ChunkShape, Codec, DataType,
+        FillValue, Path, StorageTransformer, UserAttributes, ZarrArrayMetadata,
+    },
     storage::{InMemoryStorage, MemCachingStorage},
-    ArrayIndices, ChunkKeyEncoding, ChunkPayload, ChunkShape, Codec, DataType, Dataset,
-    FillValue, Path, Storage, StorageTransformer, UserAttributes, ZarrArrayMetadata,
+    Dataset, Storage,
 };
 use itertools::Itertools;
 
@@ -161,7 +164,7 @@ ds.flush().await?;
     println!("## Adding an inline chunk");
     ds.set_chunk_ref(
         array1_path.clone(),
-        ArrayIndices(vec![0]),
+        ChunkIndices(vec![0]),
         Some(ChunkPayload::Inline("hello".into())),
     )
     .await?;
@@ -170,19 +173,19 @@ ds.flush().await?;
 ```
 ds.set_chunk(
     array1_path.clone(),
-    ArrayIndices(vec![0]),
+    ChunkIndices(vec![0]),
     Some(ChunkPayload::Inline(b"hello".into())),
 )
 .await?;
 ```
  "#,
     );
-    let chunk = ds.get_chunk_ref(&array1_path, &ArrayIndices(vec![0])).await.unwrap();
+    let chunk = ds.get_chunk_ref(&array1_path, &ChunkIndices(vec![0])).await.unwrap();
     println!("## Reading the chunk");
     println!(
         r#"
 ```
-let chunk = ds.get_chunk_ref(&array1_path, &ArrayIndices(vec![0])).await.unwrap();
+let chunk = ds.get_chunk_ref(&array1_path, &ChunkIndices(vec![0])).await.unwrap();
 => {chunk:?}
 ```
  "#
@@ -217,8 +220,8 @@ let mut ds = Dataset::update(Arc::clone(&storage), ObjectId.from("{v2_id:?}"));
     println!("## Adding a new inline chunk");
     ds.set_chunk_ref(
         array1_path.clone(),
-        ArrayIndices(vec![1]),
-        Some(icechunk::ChunkPayload::Inline("bye".into())),
+        ChunkIndices(vec![1]),
+        Some(ChunkPayload::Inline("bye".into())),
     )
     .await?;
 
@@ -227,7 +230,7 @@ let mut ds = Dataset::update(Arc::clone(&storage), ObjectId.from("{v2_id:?}"));
 ```
 ds.set_chunk(
     array1_path.clone(),
-    ArrayIndices(vec![1]),
+    ChunkIndices(vec![1]),
     Some(icechunk::ChunkPayload::Inline(b"bye".into())),
 )
 .await?;
@@ -235,10 +238,10 @@ ds.set_chunk(
  "#,
     );
 
-    let chunk = ds.get_chunk_ref(&array1_path, &ArrayIndices(vec![1])).await.unwrap();
+    let chunk = ds.get_chunk_ref(&array1_path, &ChunkIndices(vec![1])).await.unwrap();
     println!("Reading the new chunk => `{chunk:?}`");
 
-    let chunk = ds.get_chunk_ref(&array1_path, &ArrayIndices(vec![0])).await.unwrap();
+    let chunk = ds.get_chunk_ref(&array1_path, &ChunkIndices(vec![0])).await.unwrap();
     println!("Reading the old chunk => `{chunk:?}`");
 
     println!();
@@ -266,10 +269,10 @@ let ds = Dataset::update(Arc::clone(&storage), ObjectId::from("{v2_id:?}"));
 
     print_nodes(&ds).await;
 
-    let chunk = ds.get_chunk_ref(&array1_path, &ArrayIndices(vec![0])).await.unwrap();
+    let chunk = ds.get_chunk_ref(&array1_path, &ChunkIndices(vec![0])).await.unwrap();
     println!("Reading the old chunk: {chunk:?}");
 
-    let chunk = ds.get_chunk_ref(&array1_path, &ArrayIndices(vec![1])).await;
+    let chunk = ds.get_chunk_ref(&array1_path, &ChunkIndices(vec![1])).await;
     println!("Reading the new chunk: {chunk:?}");
 
     Ok(())
