+++ src/format/attributes.rs
@@ -1,0 +1,6 @@
+use arrow::array::RecordBatch;
+
+#[derive(Debug, PartialEq)]
+pub struct AttributesTable {
+    pub batch: RecordBatch,
+}
